requirements.txt
django==4.2.5
djangorestframework==3.14.0
pandas
matplotlib
seaborn
scikit-learn
nltk

